master-node/requirements.txt
fastapi
uvicorn[standard]
python-socketio
python-multipart
sqlalchemy
aiosqlite
pydantic
pydantic-settings
opencv-python
ultralytics
numpy
scipy
scikit-learn
scikit-image
pillow
aiofiles
python-dotenv
torch
torchvision

